rename: 도메인 모듈 패키지 수정
