--- domain/src/main/kotlin/com/backgu/amaker/chat/domain/ChatRoomType.kt
@@ -1,6 +1,0 @@
-package com.backgu.amaker.chat.domain
-
-enum class ChatRoomType {
-    DEFAULT,
-    CUSTOM,
-}
