--- domain/src/main/kotlin/com/backgu/amaker/chat/domain/ChatType.kt
@@ -1,8 +1,0 @@
-package com.backgu.amaker.chat.domain
-
-enum class ChatType {
-    GENERAL,
-    TASK,
-    REPLY,
-    REACTION,
-}
