--- domain/src/main/kotlin/com/backgu/amaker/common/domain/BaseTime.kt
@@ -1,8 +1,0 @@
-package com.backgu.amaker.common.domain
-
-import java.time.LocalDateTime
-
-abstract class BaseTime(
-    val createdAt: LocalDateTime = LocalDateTime.now(),
-    val updatedAt: LocalDateTime = LocalDateTime.now(),
-)
