+++ domain/src/main/kotlin/com/backgu/amaker/domain/chat/ChatRoomType.kt
@@ -1,0 +1,6 @@
+package com.backgu.amaker.domain.chat
+
+enum class ChatRoomType {
+    DEFAULT,
+    CUSTOM,
+}
