+++ domain/src/main/kotlin/com/backgu/amaker/domain/chat/ChatType.kt
@@ -1,0 +1,8 @@
+package com.backgu.amaker.domain.chat
+
+enum class ChatType {
+    GENERAL,
+    TASK,
+    REPLY,
+    REACTION,
+}
