+++ domain/src/main/kotlin/com/backgu/amaker/domain/common/BaseTime.kt
@@ -1,0 +1,8 @@
+package com.backgu.amaker.domain.common
+
+import java.time.LocalDateTime
+
+abstract class BaseTime(
+    val createdAt: LocalDateTime = LocalDateTime.now(),
+    val updatedAt: LocalDateTime = LocalDateTime.now(),
+)
