+++ domain/src/main/kotlin/com/backgu/amaker/domain/workspace/WorkspaceRole.kt
@@ -1,0 +1,9 @@
+package com.backgu.amaker.domain.workspace
+
+enum class WorkspaceRole(
+    var key: String,
+    var value: String,
+) {
+    MEMBER("ROLE_MEMBER", "Member"),
+    LEADER("ROLE_LEADER", "Leader"),
+}
