+++ domain/src/main/kotlin/com/backgu/amaker/domain/workspace/WorkspaceUserStatus.kt
@@ -1,0 +1,9 @@
+package com.backgu.amaker.domain.workspace
+
+enum class WorkspaceUserStatus(
+    var key: String,
+    var value: String,
+) {
+    PENDING("PENDING", "Pending"),
+    ACTIVE("ACTIVE", "Active"),
+}
