--- domain/src/main/kotlin/com/backgu/amaker/event/domain/EventAssignedUser.kt
@@ -1,12 +1,0 @@
-package com.backgu.amaker.event.domain
-
-import java.time.LocalDateTime
-
-class EventAssignedUser(
-    val id: Long = 0L,
-    val eventId: Long,
-    val userId: String,
-    val isFinished: Boolean = false,
-    val createdAt: LocalDateTime = LocalDateTime.now(),
-    val updatedAt: LocalDateTime = LocalDateTime.now(),
-)
