--- domain/src/main/kotlin/com/backgu/amaker/user/domain/UserRole.kt
@@ -1,10 +1,0 @@
-package com.backgu.amaker.user.domain
-
-enum class UserRole(
-    var key: String,
-    var value: String,
-) {
-    USER("ROLE_USER", "USER"),
-    MANAGER("ROLE_MANAGER", "MANAGER"),
-    ADMIN("ROLE_ADMIN", "ADMIN"),
-}
