--- domain/src/main/kotlin/com/backgu/amaker/workspace/domain/WorkspaceRole.kt
@@ -1,9 +1,0 @@
-package com.backgu.amaker.workspace.domain
-
-enum class WorkspaceRole(
-    var key: String,
-    var value: String,
-) {
-    MEMBER("ROLE_MEMBER", "Member"),
-    LEADER("ROLE_LEADER", "Leader"),
-}
