--- domain/src/main/kotlin/com/backgu/amaker/workspace/domain/WorkspaceUserStatus.kt
@@ -1,9 +1,0 @@
-package com.backgu.amaker.workspace.domain
-
-enum class WorkspaceUserStatus(
-    var key: String,
-    var value: String,
-) {
-    PENDING("PENDING", "Pending"),
-    ACTIVE("ACTIVE", "Active"),
-}
